Home page edit and delete

+++ utils/deleteOldImage.js
@@ -18,16 +18,21 @@ function deleteManyOldImage(imagePath){
 }
 
 
-function deleteSingleOldImage(imagePath){
+function deleteSingleOldImage(imagePath) {
+    if (!imagePath) {
+        console.log("Error: imagePath is undefined or empty");
+        return; // Funksiyanı dayandır
+    }
+
     const fullImagePath = path.join(imagePath);
-    fs.unlink(fullImagePath,(error)=>{
+    fs.unlink(fullImagePath, (error) => {
         if (error) {
-            console.log("Wrong prosses");
-        }else{
-            console.log("Old image delete successfull");
-            
+            console.log("Wrong process:", error);
+        } else {
+            console.log("Old image delete successful");
         }
-    })
+    });
 }
 
-module.exports = {deleteManyOldImage,deleteSingleOldImage}
+
+module.exports = {deleteManyOldImage, deleteSingleOldImage}
